requirements.txt
langgraph>=0.0.26
langchain-openai>=0.1.1
langchain-core>=0.1.30
python-dotenv>=1.0.0
httpx>=0.27.0
mcp[cli]